YUK-10503 MCT load profile peak reports can be inaccurate Changed llp_peak_report_interest to use unsigned bitfield instead of unnecessary unsigned short

+++ yukon-server/RTDB/include/dev_mct4xx.h
@@ -206,8 +206,8 @@ protected:
     struct llp_peak_report_interest_t
     {
         CtiDate end_date;
-        unsigned short range   : 10;
-        unsigned short channel :  2;
+        unsigned range   : 10;
+        unsigned channel :  2;
 
         enum States {
             Idle,
